+++ db/migrate/20141010183003_add_email_me_when_mentioned_to_users.rb
@@ -1,0 +1,5 @@
+class AddEmailMeWhenMentionedToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :email_me_when_mentioned, :boolean, null: false, default: true
+  end
+end